[MAIN-35] added user data model

+++ common/src/main/java/io/electrica/common/helper/TokenHelper.java
@@ -26,7 +26,7 @@ public class TokenHelper {
     private TokenHelper() {
     }
 
-    public static String tokenUsernameFromId(int userId) {
+    public static String tokenUsernameFromId(long userId) {
         return ID_INDICATOR + userId;
     }
